Create PersonalDeadlineLocalNotificationContentProvider
